fix: prevent exception about ``x not in list`` upon ``remove(x)`` (fixes #198)
Not sure why this is required as the code seems legit. The only case
that this will fix is the proper releasing of connectionlock upon inner
exception that would be silenced by calling code.

This was tested only on my own installation that would trigger the
exception a few hundred times, and didn't at all after this fix.

Signed-off-by: Valentin Lab <valentin.lab@kalysto.org>